refactor: #2
