+++ src/controllers/menus.controller.js
@@ -1,23 +1,26 @@
 const { View } = require('../view/view');
 const { Keyboards } = require('../keyboards/keyboard');
+const { UserSettingView } = require('../view/user-setting-view');
 
 class MenusController {
 	#userServices;
 	#postServices;
 	#commentServices;
 	#view;
+	#userSettingView;
 	#keyboard;
 	constructor(UserServices, PostServices, CommentServices) {
 		this.#userServices = UserServices;
 		this.#postServices = PostServices;
 		this.#commentServices = CommentServices;
 		this.#view = new View(UserServices, PostServices, CommentServices);
+		this.#userSettingView = new UserSettingView(UserServices, PostServices, CommentServices);
 		this.#keyboard = new Keyboards();
 	}
 
 	async mainMenu(ctx) {
 		if (ctx.message.text === ctx.i18n.t('buttons.userSetting')) {
-			this.#view.userSetting(ctx);
+			this.#userSettingView.userSetting(ctx);
 		} else if (ctx.message.text === ctx.i18n.t('buttons.userPost')) {
 			this.#view.userPost(ctx);
 		} else if (ctx.message.text === ctx.i18n.t('buttons.getAnotherUser')) {
@@ -33,7 +36,7 @@ class MenusController {
 	}
 
 	async userSettingView(ctx) {
-		this.#view.userSetting(ctx);
+		this.#userSettingView.userSetting(ctx);
 	}
 }
