+++ src/view/view.js
@@ -52,58 +52,6 @@ class View {
 		ctx.scene.enter('mainMenu');
 	}
 
-	async userSetting(ctx) {
-		ctx.reply(ctx.i18n.t('phrases.userSetting'), this.#keyboards.userSetting(ctx));
-		ctx.scene.enter('userSetting');
-	}
-
-	async userSettingName(ctx) {
-		ctx.reply(ctx.i18n.t('phrases.userSettingName'), this.#keyboards.backAndMainMenu(ctx));
-		ctx.scene.enter('userSettingName');
-	}
-
-	async newUserName(ctx, name) {
-		ctx.reply(ctx.i18n.t('phrases.newUserName').replace('$userName', name), this.#keyboards.backAndMainMenu(ctx));
-	}
-
-	async userSettingNickname(ctx) {
-		ctx.reply(ctx.i18n.t('phrases.userSettingNickname'), this.#keyboards.backAndMainMenu(ctx));
-		ctx.scene.enter('userSettingNickname');
-	}
-
-	async nicknameTaken(ctx) {
-		ctx.reply(ctx.i18n.t('phrases.busyNickname'), this.#keyboards.backAndMainMenu(ctx));
-	}
-
-	async newNickname(ctx, nickname) {
-		ctx.reply(
-			ctx.i18n.t('phrases.newUserNickname').replace('$userNickname', nickname),
-			this.#keyboards.backAndMainMenu(ctx)
-		);
-	}
-
-	async userSettingAge(ctx) {
-		ctx.reply(ctx.i18n.t('phrases.userSettingAge'), this.#keyboards.backAndMainMenu(ctx));
-		ctx.scene.enter('userSettingAge');
-	}
-
-	async enterNumber(ctx) {
-		ctx.reply(ctx.i18n.t('phrases.enterNumber'), this.#keyboards.backAndMainMenu(ctx));
-	}
-
-	async newAge(ctx, age) {
-		ctx.reply(ctx.i18n.t('phrases.newUserAge').replace('$userAge', age), this.#keyboards.backAndMainMenu(ctx));
-	}
-
-	async userSettingSex(ctx) {
-		ctx.reply(ctx.i18n.t('phrases.userSettingSex'), this.#keyboards.userSettingSex(ctx));
-		ctx.scene.enter('userSettingSex');
-	}
-
-	async newSex(ctx, sex) {
-		ctx.reply(ctx.i18n.t('phrases.newUserSex').replace('$userSex', sex));
-	}
-
 	async userPost(ctx) {
 		ctx.reply(ctx.i18n.t('phrases.userPost'), this.#keyboards.userPost(ctx));
 		ctx.scene.enter('userPost');
